+++ tests/visual/test_visual_noop.py
@@ -1,0 +1,7 @@
+import pikepdf
+
+import pdftl.api
+
+
+def test_visual_noop(six_page_rotated_pdf, assert_pdf_match):
+    assert_pdf_match(six_page_rotated_pdf)